requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
langchain-huggingface
python-dotenv
spacy
pdfminer.six 
pyresparser 
docx2txt
pip